requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
